move files to new directory
