+++ vllm/model_executor/layers/quantization/compressed_tensors/compressed_tensors_moe/compressed_tensors_moe.py
@@ -25,11 +25,6 @@ from vllm.platforms import current_platform
 logger = init_logger(__name__)
 
 
-__all__ = [
-    "CompressedTensorsMoEMethod",
-]
-
-
 class CompressedTensorsMoEMethod(FusedMoEMethodBase):
     @staticmethod
     def get_moe_method(
@@ -67,7 +62,7 @@ class CompressedTensorsMoEMethod(FusedMoEMethodBase):
         format = scheme_dict.get("format")
 
         if quant_config._is_mxfp4(weight_quant):
-            from vllm.model_executor.layers.quantization.compressed_tensors.compressed_tensors_moe_w4a4_mxfp4 import (  # noqa E501
+            from .compressed_tensors_moe.compressed_tensors_moe_w4a4_mxfp4 import (
                 CompressedTensorsW4A4Mxfp4MoEMethod,
             )
 
@@ -96,7 +91,7 @@ class CompressedTensorsMoEMethod(FusedMoEMethodBase):
                 not check_moe_marlin_supports_layer(layer, group_size)
                 or current_platform.is_rocm()
             ):
-                from vllm.model_executor.layers.quantization.compressed_tensors.compressed_tensors_moe_wna16 import (  # noqa E501
+                from .compressed_tensors_moe.compressed_tensors_moe_wna16 import (
                     CompressedTensorsWNA16MoEMethod,
                 )
 
@@ -113,7 +108,7 @@ class CompressedTensorsMoEMethod(FusedMoEMethodBase):
                     weight_quant, input_quant, layer.moe_config
                 )
             else:
-                from vllm.model_executor.layers.quantization.compressed_tensors.compressed_tensors_moe_wna16_marlin import (  # noqa E501
+                from .compressed_tensors_moe.compressed_tensors_moe_wna16_marlin import (  # noqa: E501
                     CompressedTensorsWNA16MarlinMoEMethod,
                 )
 
@@ -122,7 +117,7 @@ class CompressedTensorsMoEMethod(FusedMoEMethodBase):
                     weight_quant, input_quant, layer.moe_config
                 )
         elif quant_config._is_nvfp4_format(weight_quant):
-            from vllm.model_executor.layers.quantization.compressed_tensors.compressed_tensors_moe_w4a4_nvfp4 import (  # noqa E501
+            from .compressed_tensors_moe.compressed_tensors_moe_w4a4_nvfp4 import (
                 CompressedTensorsW4A4Nvfp4MoEMethod,
             )
 
@@ -142,7 +137,7 @@ class CompressedTensorsMoEMethod(FusedMoEMethodBase):
             or quant_config._is_fp8_w8a8_sm100(weight_quant, input_quant)
             or quant_config._is_fp8_w8a8(weight_quant, input_quant)
         ):
-            from vllm.model_executor.layers.quantization.compressed_tensors.compressed_tensors_moe_w8a8_fp8 import (  # noqa E501
+            from .compressed_tensors_moe.compressed_tensors_moe_w8a8_fp8 import (
                 CompressedTensorsW8A8Fp8MoEMethod,
             )
 
@@ -150,7 +145,7 @@ class CompressedTensorsMoEMethod(FusedMoEMethodBase):
                 weight_quant, input_quant, layer.moe_config
             )
         elif quant_config._is_dynamic_token_w8a8(weight_quant, input_quant):
-            from vllm.model_executor.layers.quantization.compressed_tensors.compressed_tensors_moe_w8a8_int8 import (  # noqa E501
+            from .compressed_tensors_moe.compressed_tensors_moe_w8a8_int8 import (
                 CompressedTensorsW8A8Int8MoEMethod,
             )
 
@@ -158,7 +153,7 @@ class CompressedTensorsMoEMethod(FusedMoEMethodBase):
                 weight_quant, input_quant, layer.moe_config
             )
         elif quant_config._is_fp8_w4a8_sm90(weight_quant, input_quant):
-            from vllm.model_executor.layers.quantization.compressed_tensors.compressed_tensors_moe_w4a8_fp8 import (  # noqa E501
+            from .compressed_tensors_moe.compressed_tensors_moe_w4a8_fp8 import (
                 CompressedTensorsW4A8Fp8MoEMethod,
             )
 
@@ -167,7 +162,7 @@ class CompressedTensorsMoEMethod(FusedMoEMethodBase):
                 weight_quant, input_quant, layer.moe_config
             )
         elif quant_config._is_dynamic_token_w4a8_int(weight_quant, input_quant):
-            from vllm.model_executor.layers.quantization.compressed_tensors.compressed_tensors_moe_w4a8_int8 import (  # noqa E501
+            from .compressed_tensors_moe.compressed_tensors_moe_w4a8_int8 import (
                 CompressedTensorsW4A8Int8MoEMethod,
             )
